@@ -4,59 +4,62 @@
   "scripts": {
     "ng": "ng",
     "start": "ng serve",
-    "stats": "webpack-bundle-analyzer dist/web-application/stats.json",
     "build": "ng build",
     "test": "ng test",
     "lint": "ng lint",
-    "e2e": "ng e2e"
+    "e2e": "ng e2e",
+    "stats": "webpack-bundle-analyzer dist/web-application/stats.json",
+    "lib": "ng build --project sc-common",
+    "dev": "npm run lib && ng serve"
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^6.0.3",
-    "@angular/cdk": "^6.2.0",
-    "@angular/common": "^6.0.3",
-    "@angular/compiler": "^6.0.3",
-    "@angular/core": "^6.0.3",
-    "@angular/flex-layout": "^6.0.0-beta.17",
-    "@angular/forms": "^6.0.3",
-    "@angular/http": "^6.0.3",
-    "@angular/material": "^6.4.2",
-    "@angular/platform-browser": "^6.0.3",
-    "@angular/platform-browser-dynamic": "^6.0.3",
-    "@angular/pwa": "^0.6.8",
-    "@angular/router": "^6.0.3",
-    "@angular/service-worker": "^6.0.3",
+    "@angular/animations": "^7.0.0",
+    "@angular/cdk": "^7.0.0",
+    "@angular/common": "^7.0.0",
+    "@angular/compiler": "^7.0.0",
+    "@angular/core": "^7.0.0",
+    "@angular/flex-layout": "^7.0.0-beta.19",
+    "@angular/forms": "^7.0.0",
+    "@angular/http": "^7.0.0",
+    "@angular/material": "^7.0.0",
+    "@angular/platform-browser": "^7.0.0",
+    "@angular/platform-browser-dynamic": "^7.0.0",
+    "@angular/pwa": "^0.8.4",
+    "@angular/router": "^7.0.0",
+    "@angular/service-worker": "^7.0.0",
     "core-js": "^2.5.4",
-    "ngx-socket-io": "^1.1.1",
-    "rxjs": "^6.0.0",
-    "rxjs-compat": "^6.2.2",
-    "zone.js": "^0.8.26"
+    "rxjs": "~6.3.3",
+    "rxjs-compat": "^6.3.3",
+    "zone.js": "~0.8.26"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.6.8",
-    "@angular/cli": "~6.0.8",
-    "@angular/compiler-cli": "^6.0.3",
-    "@angular/language-service": "^6.0.3",
+    "@angular-devkit/build-angular": "~0.8.0",
+    "@angular-devkit/build-ng-packagr": "~0.8.0",
+    "@angular/cli": "^7.0.3",
+    "@angular/compiler-cli": "^7.0.0",
+    "@angular/language-service": "^7.0.0",
     "@compodoc/gulp-compodoc": "0.0.9",
-    "@types/jasmine": "~2.8.6",
+    "@types/jasmine": "~2.8.8",
     "@types/jasminewd2": "~2.0.3",
     "@types/node": "~8.9.4",
-    "codelyzer": "~4.2.1",
+    "codelyzer": "~4.3.0",
     "gulp": "^3.9.1",
-    "gulp-load-plugins": "^1.5.0",
-    "gulp-responsive": "^2.11.0",
     "gulp-typescript": "^5.0.0-alpha.3",
     "gulp-util": "^3.0.8",
     "jasmine-core": "~2.99.1",
     "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~1.7.1",
+    "karma": "~3.0.0",
     "karma-chrome-launcher": "~2.2.0",
-    "karma-coverage-istanbul-reporter": "~2.0.0",
-    "karma-jasmine": "~1.1.1",
+    "karma-coverage-istanbul-reporter": "~2.0.1",
+    "karma-jasmine": "~1.1.2",
     "karma-jasmine-html-reporter": "^0.2.2",
-    "protractor": "~5.3.0",
-    "ts-node": "~5.0.1",
-    "tslint": "~5.9.1",
-    "typescript": "~2.7.2"
+    "ng-packagr": "^4.4.0",
+    "protractor": "~5.4.0",
+    "ts-node": "~7.0.0",
+    "tsickle": "^0.33.0",
+    "tslib": "^1.9.3",
+    "tslint": "~5.11.0",
+    "typescript": "^3.1.3"
   }
 }
