@@ -5,13 +5,6 @@ import { FormsModule } from '@angular/forms';
 import { HttpClientModule } from '@angular/common/http';
 import { RouterModule } from '@angular/router';
 
-import { LayoutModule } from '@angular/cdk/layout';
-import { SocketIoModule } from 'ngx-socket-io';
-
-import { FooterComponent } from './components/footer/footer.component';
-import { NavComponent } from './components/nav/nav.component';
-import { SharedModule } from '../shared/shared.module';
-
 /**
  * Module which includes Application's main utils and singleton services. This module is only imported by AppModule.
  *
@@ -25,24 +18,14 @@ import { SharedModule } from '../shared/shared.module';
     FlexLayoutModule,
     FormsModule,
     HttpClientModule,
-    RouterModule,
-    SocketIoModule,
-    LayoutModule,
-    SharedModule
+    RouterModule
   ],
-  declarations: [
-    FooterComponent,
-    NavComponent
-  ],
+  declarations: [],
   exports: [
     CommonModule,
     FlexLayoutModule,
-    FooterComponent,
     FormsModule,
-    HttpClientModule,
-    NavComponent,
-    SharedModule,
-    SocketIoModule
-  ],
+    HttpClientModule
+  ]
 })
 export class CoreModule { }
